Sort SOAP response elements
Some parsers are picky about the order of elements in SOAP responses. It
is recommended to follow the order given in the spec/wsdl. This change
replaces the map used to pass the values back to the handler with a
slice of 2-arrays, such that imposing an order is possible, and sorts
the keys in the correct order.

Verified with UPnP Monkey on Android, which before said:

> ServerHandler: failure: org.fourthline.cling.model.action.ActionException: Current state of service prevents invoking that action. Error reading response message. Can't transform message payload: org.fourthline.cling.model.action.ActionException: The argument value is invalid. Wrong order of arguments, expected 'Result' not: TotalMatches..