use goreleaser.yml
pass path to config file:
this was annoying to figure out. there's no manpage or other docs for this I
found the option at
https://github.com/goreleaser/goreleaser/discussions/2644
It's also annoying that I need this file at all; I feel that I should be able
to pass the options to goreleaser directly, but don't know how to.